+++ index.html
@@ -24,6 +24,9 @@
          src="https://kit.fontawesome.com/ce1e161f8c.js"
          crossorigin="anonymous"
       ></script>
+
+      <script src="https://d3js.org/d3.v4.js"></script>
+      <script src="https://cdn.jsdelivr.net/gh/holtzy/D3-graph-gallery@master/LIB/d3.layout.cloud.js"></script>
    </head>
    <body>
       <div id="header">
@@ -72,9 +75,9 @@
             <div id="scatter-histogram-charts-container"></div>
          </div>
 
-         <div class="chart-container" id="word-search-container">
-            <div class="heading">Word Search</div>
-            <div id="word-search-container"></div>
+         <div class="chart-container" id="word-cloud-container">
+            <div class="heading">Word Cloud</div>
+            <div id="word-cloud-diagram-container"></div>
          </div>
 
       </div>
