+++ js/charts/index.mjs
@@ -1,4 +1,5 @@
 export * from './bar.js';
 export * from './histogram.js';
 export * from './leafletMap.js';
+export * from './word-cloud.js';
